requirements.txt
django==3.1.7
channels
django-allauth
django-crispy-forms
django-taggit
pillow